refactor: migrate code from js to ts
